refactor: implement dry in gpg
